fix: logic and style separation
